server/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Тимур Вахитов <vt@greentechs.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.57.0"
grpcio-tools = "^1.57.0"
setuptools = "^68.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
